requirements.txt
fastapi
uvicorn
pydantic
aiosmtplib
pytest
pytest-asynciodoc